// src/modules/shared/models/breakingBand.interface.ts
export interface BreakingBadCharacter {
  char_id: number;
  name: string;
  birthday: string;
  occupation: string[];
  img: string;
  status: string;
  nickname: string;
  appearance: number[];
  portrayed: string;
  category: string;
  better_call_saul_appearance: number[];
}

export interface Character {
  char_id: number;
  name: string;
  nickname: string;
  status: string;
}